requirements.txt
Django==2.2.1
django-ckeditor==5.7.0
django-haystack==2.8.1
django-js-asset==1.2.2
jieba==0.39
mysqlclient==1.4.2.post1
Pillow==5.3.0
pytz==2019.1
sqlparse==0.3.0
Whoosh==2.7.4
